x/bears/types/message_set_apiary_house_for_bee_test.go
package types

import (
	"testing"

	"github.com/MonetaToday/HoneyWood/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgSetApiaryHouseForBee_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgSetApiaryHouseForBee
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgSetApiaryHouseForBee{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgSetApiaryHouseForBee{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
